got tabledraggable working with everything extracted to a context
