+++ src/index.tsx
@@ -1,3 +1,3 @@
-export { default as DragTable, DragTableProps } from './DragTable'
 export { default, TableProps } from './Table'
+export { default as DragTable, DragTableProps } from './TableDraggable'
 export { Item, BetterItem, FunctionalItem } from './lib'
